requirements.txt
wget
tgcrypto
ffmpeg-python
py-tgcalls==0.9.7
pyrogram==2.0.106
youtube-search
youtube-search-python
yt-dlp
python-dotenv
dnspython
gitpython
asgiref
aiofiles
aiohttp
requests
pillow
heroku3
motor
psutil
future
wget
speedtest-cli
pymongo[srv]
pyromod~=1.5
googletrans==4.0.0-rc1
telethon